server/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.1.0.post1

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9  # PostgreSQL用
python-dotenv==1.0.0

# OCR
paddleocr==2.7.0
paddlepaddle==2.5.2

# AI/ML
langchain==0.0.335
openai==1.2.4

# Utilities
python-jose==3.3.0  # JWT
passlib==1.7.4  # パスワードハッシュ
bcrypt==4.0.1  # パスワードハッシュ
uuid==1.30
pillow==10.1.0  # 画像処理

# Testing
pytest==7.4.3
httpx==0.25.1
